feat(seletor): mover para pasta especifica e tornar construtor publico
